@@ -6,7 +6,7 @@
   "license": "AGPL-3.0",
   "repository": {},
   "engines": {
-    "node": ">=18"
+    "node": ">=20.9.0"
   },
   "scripts": {
     "build": "run-s -l build:angular1 build:angular17",
@@ -26,24 +26,24 @@
   "keywords": [],
   "author": "",
   "dependencies": {
-    "@angular/animations": "^17.0.3",
-    "@angular/cdk": "^17.0.1",
-    "@angular/cli": "^17.0.1",
-    "@angular/common": "^17.0.3",
-    "@angular/compiler": "^17.0.3",
-    "@angular/core": "^17.0.3",
+    "@angular/animations": "^17.3.5",
+    "@angular/cdk": "^17.2.1",
+    "@angular/cli": "^17.3.5",
+    "@angular/common": "^17.3.5",
+    "@angular/compiler": "^17.3.5",
+    "@angular/core": "^17.3.5",
     "@angular/flex-layout": "^15.0.0-beta.42",
-    "@angular/forms": "^17.0.3",
-    "@angular/material": "^17.0.1",
-    "@angular/material-moment-adapter": "^17.0.1",
-    "@angular/platform-browser": "^17.0.3",
-    "@angular/platform-browser-dynamic": "^17.0.3",
-    "@angular/router": "^17.0.3",
-    "@angular/service-worker": "^17.0.3",
-    "@angular/upgrade": "^17.0.3",
+    "@angular/forms": "^17.3.5",
+    "@angular/material": "^17.2.1",
+    "@angular/material-moment-adapter": "^17.2.1",
+    "@angular/platform-browser": "^17.3.5",
+    "@angular/platform-browser-dynamic": "^17.3.5",
+    "@angular/router": "^17.3.5",
+    "@angular/service-worker": "^17.3.5",
+    "@angular/upgrade": "^17.3.5",
     "@ctrl/ngx-emoji-mart": "^9.2.0",
-    "@uirouter/angular": "^12.0",
-    "@uirouter/angular-hybrid": "^16.0",
+    "@uirouter/angular": "^13.0",
+    "@uirouter/angular-hybrid": "^17.1.0",
     "@uirouter/angularjs": "^1.0.30",
     "@uirouter/core": "^6.1.0",
     "@uirouter/rx": "^1.0.0",
@@ -75,31 +75,30 @@
     "jquery": "2.1.4",
     "lodash": "~4.17",
     "lottie-web": "^5",
-    "marked": "^10.0",
-    "moment": "^2.29.1",
+    "marked": "^11.1.0",
+    "moment": "^2.29.4",
     "ng-csv": "0.2.3",
     "ng-file-upload": "~5.0.9",
     "ng2-pdf-viewer": "^10.0",
     "ngx-bootstrap": "^6.1.0",
-    "ngx-entity-service": "^0.0.37",
+    "ngx-entity-service": "^0.0.38",
     "ngx-lottie": "^10.0.0",
     "nvd3": "1.8.6",
     "rxjs": "~7.4.0",
-    "showdown": "1.3.0",
-    "ts-md5": "^1.2.7",
-    "tslib": "^2.3.1",
+    "ts-md5": "^1.3.1",
+    "tslib": "^2.6.2",
     "underscore.string": "2.3.3",
     "zone.js": "~0.14"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.1",
-    "@angular-eslint/builder": "^17.0.0",
-    "@angular-eslint/eslint-plugin": "^17.0.0",
-    "@angular-eslint/eslint-plugin-template": "^17.0.0",
-    "@angular-eslint/schematics": "^17.0.0",
-    "@angular-eslint/template-parser": "^17.0.0",
-    "@angular/compiler-cli": "^17.0.3",
-    "@angular/language-service": "^17.0.3",
+    "@angular-devkit/build-angular": "^17.3.5",
+    "@angular-eslint/builder": "^17.1.1",
+    "@angular-eslint/eslint-plugin": "^17.1.1",
+    "@angular-eslint/eslint-plugin-template": "^17.1.1",
+    "@angular-eslint/schematics": "^17.1.1",
+    "@angular-eslint/template-parser": "^17.1.1",
+    "@angular/compiler-cli": "^17.3.5",
+    "@angular/language-service": "^17.3.5",
     "@commitlint/cli": "^16.0.1",
     "@commitlint/config-conventional": "^17",
     "@types/angular": "1.5.11",
@@ -109,7 +108,7 @@
     "@types/jasmine": "~3.9.1",
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "^4.14.115",
-    "@types/node": "^18.0",
+    "@types/node": "^20.9.0",
     "@typescript-eslint/eslint-plugin": "^6.10.0",
     "@typescript-eslint/parser": "^6.10.0",
     "autoprefixer": "~6",
