@@ -17,7 +17,7 @@ export class MarkedPipe implements PipeTransform {
 
   transform(value: string, ...args: any[]): string {
     if (value && value.length > 0) {
-      return marked.parse(value.replaceAll(/\r\n|\r|\n/g, '<br />'));
+      return <string>marked.parse(value.replaceAll(/\r\n|\r|\n/g, '<br />'), {async: false});
     }
     return value;
   }
